pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "encryption-and-digital-signatures-using-python"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
pycryptodome = "^3.9"
[tool.poetry.dev-dependencies]
